requirements.txt
# --------- pytorch --------- #
torch==2.6
torchvision
tabulate # torch.compile prerequisite
scipy # torch.compile prerequisite
numpy # torch.compile prerequisite
lightning # pytorch lightning for training

# --------- huggingface --------- #
datasets
transformers
transformers[torch]

# --------- hydra --------- #
hydra-core
hydra-colorlog
hydra-optuna-sweeper
omegaconf

# --------- loggers --------- #
wandb
torch_tb_profiler

# --------- others --------- #
pyrootutils # project root setup
pre-commit<4 # managing and maintaining pre-commit hooks. (v4.0.1+ breaks docformatter)
rich # Render rich text, tables, progress bars, syntax highlighting, markdown etc to the terminal
pytest # simple powerful testing with Python
sh; platform_system != 'Windows'
tqdm # Fast, Extensible Progress Meter
pynvml # Python utilities for the NVIDIA Management Library
chess # A chess library with move generation and validation, etc.
python-dotenv # Read key-value pairs from a .env file and set them as environment variables
# --------- visualization --------- #
seaborn
scikit_learn
matplotlib
ipython
